fix armory

+++ server/core-api/Data/World/CustomItemUpgrades.cs
@@ -6,7 +6,7 @@
         public byte StatType { get; set; }
         public int StatValue { get; set; }
         public byte StatLevel { get; set; }
-        public byte RequiredLevel { get; set; }
+        public int RequiredLevel { get; set; }
         public int ConditionId { get; set; }
         public string AllowableInvTypes { get; set; }
     }
